+++ pkg/app/contextmenu_wasm.go
@@ -1,5 +1,3 @@
-// +build wasm
-
 package app
 
 import (
